feat: implement signup wizard with onboarding steps and infinite scroll directive
- Added a new signup wizard component with a multi-step onboarding process.
- Integrated various input types including sliders, date pickers, and text areas.
- Implemented validation for user inputs and asynchronous checks for username and email uniqueness.
- Created a new infinite scroll directive to enhance data loading capabilities in lists.
- Styled the signup wizard using SCSS with responsive design considerations and dark theme support.

+++ backend/src/main/java/com/gestioneventos/cofira/entities/UserProfile.java
@@ -54,7 +54,7 @@ public class UserProfile {
     @Enumerated(EnumType.STRING)
     private WorkType workType;
 
-    private Integer sleepHoursAverage;
+    private Double sleepHoursAverage;
 
     // Goals & Training
     @Enumerated(EnumType.STRING)
